SQL for groups containing meters and other groups
groups.sql adds tables and views for handling a directed acyclic graph of "groups", each containing zero or more meters. Meters are organized in a multitree structure that lets the groups_deep_meters view mark meters as "shadowed" if they cover an energy reading that is also covered by a more general meter in the current group.