Adds database test framework and deletes old tests
Run the tests with npm run test.

This adds configuration options for a test database. It also adds an example
test on that database that creates and retrieves a meter.

The database is specified using the DB_TEST_DATABASE environment variable.
DB_TEST_USER, DB_TEST_HOST, etc. can be optionally specified, but production
setttings will be used if they are not found.

This also deletes the old tests that were sitting in the src/server/tests directory.
They aren't really tests and were mostly a sandbox for very early ideas.