Changes readings to an interval. Creates compressed_readings function
This modifies the readings table and associated javascript so that a reading
occurs over a time interval instead of at a point in time. This makes sense
because a reading is actually the amount of energy consumed over a given
interval, and therefore the information in the database was previously
insufficient for truly representing the readings.

The get_compressed_readings function exploits this to compress readings for
a meter down to a smaller number of readings for transmission to a client and
therefore easier graphing.