Fixes bugs and adds tests in compression code
This fixes bugs where the number of compressed points was always 1 larger than specified,
and where specifying a date range during compression excluded points that directly touched
the edge of that date range.

It also switches to a generator-style way of running queries in tests, and
it adds optional arguments to methods in Meter.js and Reading.js to use a
specific database connection, such as one from a database transaction or task.