Refactor tests
The various database tests were starting to get unruly, so
this commit creates a new test subdirectory called db with
files containing tests for different parts of the database.

It also separates code to patch the config and recreate the
database into a common.js file, which gets required by the
other files.