Refactor redux state and add multiline charts
This refactors redux's state shape to be more normalized and splits the actions
and reducers across multiple files.

It also changes the way that the line chart works so that it interfaces with
highcharts without recreating the chart. The new line chart is currently in
`ReduxLineChartComponent` instead of `LineChartComponent` because
the transition isn't complete yet. Likewise for `NewLineChartContainer`
 and `LineChartContainer`. Bar charts are not yet updated and therefore
 currently broken.

Finally, it adds a line in index.jsx to enable redux dev tools if they're
installed in the client browser.