Refactor start/end timestamps to common TimeInterval class
startTimestamp and endTimestamp values were being passed all over the place in
the client, and even sketchily sent to the server. This creates a common TimeInterval
class that encapsulates the concept of a time interval and makes it available to both
the client and the server. TimeInterval also handles its own serialization and
deserialization when moving between client and server.