Refactor and make major additions to groups UI.
Make major changes to the way groups are stored in state.
Add `childGroups` and `childMeters` to each group in `byGroupID`
as arrays storing the IDs of child meters and groups.
Also add an `isFetching` flag to each group to indicate whether
we are currently fetching the children of a group.
Add `selectedMeters` and `selectedGroups` to the groups
state to store which meters and groups are selected as far as the
groups UI is concerned. These are currently unused.

To make streamline fetching the children of a group, change the
files `get_immediate_groups_by_group_id.sql` and
`get_immediate_meters_by_group_id.sql` in `sql/group` to return only
the IDs of the children, rather than objects, and change functions in
`models/Group.js` to reflect this.

Add actions in `actions/group.js` to handle fetching the children
of groups, and alter the actions to fetch groups details to be
correct.

Add reducers in `reducers/groups.js` to corresponding to the new actions,
and correct reducers that handle fetching group details.

Add `ChildMeterBoxContainer` and `ChildGroupBoxContainer` and convert the
corresponding components to class components so that they will be able to
update the selected meters and groups in state.

Changes NOT directly related to groups UI:

To avoid confusion, change the names of actions that get information on
 a particular meter or group from `FETCH...DATA` to `FETCH...DETAILS`.
 And change the names of reducers, etc. to match. This accounts for all
 the changes to `actions/meters.js`, `reducers/meters.js` , and
 `reducers/readings.js`.

 In `updateCron.bash` remove the `--harmony-async-await` flag
 because it is no longer necessary.

The following issues persist:
`ChildGroupBoxComponent` displays the groups in an unordered list so it
is not possible to select them. Selecting meters in
`ChildMeterBoxComponent` generates an error.
Switching to the edit page and back removes part of the state, but only
sometimes and it's not clear why.
The display is ugly as sin.