Change how we control what attributes of groups are sent to the client.
Move control over what data is sent to the client (in groups) from the
sql files to routes.

Use a function in `routes/groups.js` to ensure that only group ID and
name are sent to the client (in case we add more fields later).

Change the `get_immediate_children` queries to return all data on the
child because the table only stores ID to begin with.