Add version adornments.
The file "common/version.js" should be updated in the LAST commit before
a tag is made for that version. The contents of the created version
object are displayed on the app and are available as JSON in the API,
from /api/version, in case we ever want to make mobile apps or support
very long-running displays (e.g. the "wall graph" use case).