Fix tests destroying the database.
This involved changing all imports of the db global to imports
of a getDB function which is a relatively transparent wrapper
over that global. It does, however, check for config changes and
enforce a total transition to any newly configured DB.

Performance impact should be minimial as that code will mostly be
skipped and the tracing JIT should inline the function fairly
rapidly.