UI for meters toggling; display only
Provides a tabular view of all meters in the system, with all associated
information (including toggles). Users who are not logged in see only
meters with the Displayable flag on, regardless of what view is in use,
and the server does not send IP or meter type in the absence of a token.