Make docker-compose up more useful.
Docker-compose now installs OED by default, so that most developers
need only run `docker-compose up` to get started. In production, the
OED_PRODUCTION variable needs to be changed in the docker-compose.yml
file, but other than that, all things are the same.

This is accomplished by making installOED.sh also start OED. It is
idempotent as long as the mode is not changed, since it now checks to
see if node_modules/ exists, so it's fine to run it on start.