update package-lock.json
The accomplishes a copuple of things:
- The changes Nicole made caused some packages to use SHA1 not SHA512.
This is a known issue and this puts many at SHA 512.
- We had many outdates and vulnerable packages which have been updated.
Note there are still 20 packages that need to be looked at.

The only known issue that some developer may see is that the upgrades
need a newer TypeScript (believed to be at least 3.5)
or they will get errors about Omit
but OED still runs.
It is also noted that there are many continuing lint errors that will
be addressed (not new to these changes).