Allow more time for mocha tests to complete
When used with Docker, the mocha tests have been timing out on database
access. The timeout was increased to 15 sec to try to avoid this.
It removes the failed tests that were happening at seemingly random
places. Still need to see if helps with Travis.
This is really a hack as the tests should not take this long.
We need to determine if there is an underlying issue with the connection
mechanism or the test setup.