Fix TypeScript issues
The upgrade in packages on development gave an error on Omit.
We needed to get to TypeScript 3.5+ and this gets us to 3.5.3.
Note going to 3.6+ generated two other issues so not
doing (current 3.8.3 same).
These will become a GitHub issue.
RouteComponent.tsx had a new TypeScript error from the upgrade.
This is fixed in a klugy way but it seems to work fine.
This is related to previous change and in GitHub issue #444.
That issue will be updated and a TODO left in code to check this out.
package-lock.json was updated for this. I note that the one in
development had a syntax error that was corrected.
There are also seemingly minor upgrades to packages with this.