update minimist & install
Fixes security advisory  GHSA-vh95-rmgr-6w4m.
Note geojson-rewind used a vulnerable version but it has not been
updated in several years so I manually changed the version to allow
for later versions.

npm i changes package-lock.json if online info differs. To overcome this
I modified the install script to use npm ci. This causes node_modules to
be removed for each install. It is reasonably fast but still slower.
To mitigate this a new option (--keep_node_modules) was
added to skip the npm install. Note that if the lock file becomes out of
date then npm ci will error and the install fail. This addresses issue
#481 but it was decided that checking modification dates for all
subdirectories might be slower and using a standard solution was better.