make install handle DB errors better
The install script continued even when createdb errors occurred.

createDB.js now returns an exit code so know its status.

installOED.sh now considers createDB exit code. If there is an error
that is not known and recoverable then it exits with error code 3.
You can override with new argument of --continue_on_db_error.
This can be useful if you want to use the web container without DB
when install will not work.
Script outputs more information on what it is doing.
Script consistently uses printf instead of echo.
Space indenting removed for tabs.

Note there are debug statements present that will be removed later.

Also note that createUser should probably work the same way but
it seems to require a change in how it deals with errors so this was
delayed. This seems less important because DB errors seem to
always happen first on createDB and then createUser runs okay.