skip empty arguments
The current docker-compose.yml expects an environment variable.
If this is not set then it passes a blank value. The install script was
failing when this happened so this fixes this.

Note that a better fix would be to stop passing the blank
argument. This has not been figured out yet (if can be done).