make createUser more resilient, fix install bug
This tried to make createUser closer to the new createDB.
The SQL avoids an error if the user exists.
createUser now stops on errors since none should happen.
A new, optional parameer to terminateReadline allows code to supply
an exit status value where default is old value of 0.
The install script for createDB retries fixes error where it stopped on
last attempt even if it worked. It also stops if createUser fails.