fix obvius bug & upgrade pg-promise
The upgrade of pg-promise is needed per discussion in a recent PR.
This does not completely fix the issues but allows current developers to
work with this. This will be completely resolved once PR is merged into
this branch (after approved).

The upgrade found a mistake I made in DB usage. You should not be doing
a dropConnection() as this does a stopDB() that invalidates DB
connection. pg-promise now detects and errors in this case.
Thus, they were removed.