add better obvius test data
The new test files have several days of data that are hourly.
This means OED can graph the data after refreshing views because
it has multiple days. The switch to hourly data fixes a temporary issue
that the the code assumes each point spans one hour (this needs to be
changed as we merge in the new pipeline code).

It is recommended to look at meter  number 4 as it has the best data.
To see the data reasonably, you need to move the right endpoint and
redraw (we need to look into this issue elsewhere as not related to
obvius). You will lose one day of the three days due to how OED
currently graphs data. Finally, you need to be logged in as admin to see
the meters since display is disabled when obvius meters are added.
You can change this for desired meter(s) on the meter page as admin.