fix moment test errors
The upgrade of pg-promise changed the names for CTF
(Custom Type Formatting) and these were fixed.
Note this also fixed the 6 failing tests in issue: test error with
pg-promise upgrade #527.
This was fixed yesterday in PR #502 but is now being done here since
this addresses all the testing errors (I hope).