set Moment's global localization based on db
Previously Moment's global localization was only set when the preferred
language was changed, causing the need to toggle between languages for
Moment to use a non-default localization on startup.

Now the global localization is also updated during fetchPreferences(),
ensuring that the preferred localization specified in the database at
state.admin.defaultLanguage is used on initial load.