Internationalize the downloadRawCSV() function
When downloading raw data, the timestamps do not ever pass through a
Moment format, which means that the global localization changes did not
affect this process.

Instead, the timestamps were default Date objects.
As such, an Intl.DateTimeFormat object is used to format the timestamp
into a localized string.

However, this likely can be simplified by converting the timestamps to
Moments instead, so that it will function like the rest of the
localization changes.