Use Moment for localization for raw data exports
Previously, the timestamps used for exporting raw data were never
formatted through Moment, so the localization changes that affected the
rest of the places that use Moment didn't affect this specific process.

Now the timestamps are passed through Moment so that the global
localization we tell Moment to use elsewhere is what controls the
localization process.