trust Plotly to find locale given our locale name
The code required to manually check and verify that a correct locale
name is used to set the `locale` property is not scalable for a greater
number of localizations - a new else would be required for all
containers that used a PlotlyChart. Instead, we rely on the Locales
module defined in Locales.ts elsewhere to be a sufficient collection of
potential localizations and we rely on the way we name our locales to
be close enough to Plotly's for them to know which localization we refer
to via state.admin.defaultLanguage.

This change means that no further modifications are necessary for all
the containers that use PlotlyCharts to use localizations that are added
later. However, any new containers will need to emulate this design.