map rotation and other map stuff
- The rotation of maps now seems to work fine. This required a lot of
changes to how the map was rotated/shifted and completely changed
how the check occurs for if meter is on the map.
- The circle size calculation was fixed.
- The calibration error is now the relative error. It was not correct before.
- Many comments were added.
- Note the map angle is hard coded and needs to come
from the DB. For now set to 0 so no rotation.
- Note the max circle size is hard coded and should come from admin
at some point.
- Note this will not work if you cross the -180/180 longitude but this
is not as likely. It can happen in a small part of northeast Russia and
Taveui Island and islands on the same longitude that may be nearby.
I am not going to deal with this unlikely case until needed.