remove dropConnection
The new version of pgp does not require us to drop connections.
OED recovers by doing a new connection but that is slower.
Previously, many of these calls were removed but some remained.
This removes most of the remaining ones except during testing where
we want to reset the DB each time.
In a couple of files only the import remained and was removed.