Add role into redux store
We did this by copying and modifying the way that the version is stored
in the redux store. We include both role and email as the 'profile' as
it seemed natural to include more than just the role in the store. This
is also just in case need the email in the future, perhaps to send
emails to the logged in user.