Remove instances of hasToken
We remove instances of hasToken to remove the import and because the
its use in those areas is to only allow authenticated admins to access
certain information. The check that we had added serves that
purpose since with user roles, simply checking for a token is not enough
to ensure that the user is an admin. Since we had forced a direct
relationship between having a token and storing the user's profile into
the redux store, checking that we have a token is redudant.