send back log output when do curl
- Messages went to the console when doing CSV upload for readings but
were not returned as part of requst. Now they come back as HTML.
This will make them look good if in a web browser or each to format
by putting into one.
- Failure is not correctly shown for any dropped reading whereas
before it was only if an exception occurred.
- It is possible that some warning/message were lost before when
another error occurred but will certainly show now.
- Values from the user are not html escaped but it is assumed done
before getting to the process step. This was now noticed but left.
- Note this does not fix what is returned for web page CSV upload nor
does it address messages for meter CSV upload.
- It also does not address other calls to the modifed methods.
This likely why one Metasys test is failing. Fix soon.