remove Metasys code
The new pipeline & CSV input offers all the input options of the old
Metasys code. Maintaining this older code no longer made sense so
it has been remove along with the associated tests. Note the old version
and the newer one when the pipeline was first begun are removed.
The one item lost is the fact that the code edited the incoming data to
remove the units on the reading. This can easily be done by a site or
another layer of code that feeds into the new pipeline.
Note the Metasys type meter is left for historical reasons.