remove CSV stream & stream; fix Obvius issue
Stream input for CSV is now well out of date given the new pipeline.
Time tests show that with regular CSV input loading it all into memory
works fairly quickly. If you have a non-server with less memory where
the amount available for JS is less than about 2GB then a file too much
over 6MB/100k points may run out of memory. This can be fixed by
increasing the memory available to JS/Node. Another fix is to split
the file into sections of smaller size. This is not expected to
happen on a live server with reasonable memory/setting. Given all of
this, it was decided to remove the stream option rather that upgrade it.

While checking for stream in code, found a number from the old meter and
reading drop zone that was replaced with the new one. All that code for
web pages, routes, translations, etc. were deleted. Also found two
stream package that are not used (one seems old issue).

Note the changes pass all but one test but have yet to be
run on live MAMAC
meters (that used to use stream but not really necessary since come as
1440 points at a time). This will be done before the PR is updated.
The failing test is believed to be caused by how errors are handled
in the new pipeline. Getting all the calling functions to correctly
handle new return values should fix it and will be done soon.

package-lock.json needs updating for package changes. Will do this
later when online.

The Obvius example upload was failing because identifier was made unique
in DB. Changed so identifier appends unique ID.