add meter name to message + other
- If the pipeline is called many times quickly, it can be hard to tell
which meter is involved. Thus, the meter name was added to
all the messages.
- Obvius log file tests failing because 1970-01-01 as the date
and this is the same as moment(0) that is used on meter when
there is no data. Also, removed overlapping test since this was
when assumed each reading was 1 hour. Finally, removed blank
line for synthetic CSV data because lead to null reading.
- 1970 is used in a other places but none seem to involve adding
readings to DB so probably okay.
- Fixed the issue with multiple return values from appendMsgTotal
so don't need different named variables.
- Fixed some check/lint issues.
- Updated expected output for new one.