add other meter type
We OED's generalization of adding meters/readings, we are beyond the
meter types known. Thus, an "other" type has been added. It is used when
a meter is automatically created but the type is not know. A site can
also use this type.
A few tests were changed to use this type to be sure it works.