make DB start/end timestamp be moment(0)
Originally we planned to user Jan 1, 0001 as the default start/end time
on a meter. The code uses moment(0) which is 1970-01-01 00:00:00.
This makes the DB consistent with the code even though it is unlikely
to ever see the DB default because the code sets a default value on
meter creation.