fix up migrations
During the last couple of new tagged versions of OED there was not
careful testing that the DB migrations worked properly. This hopefully
fixes all the issues:
- Put in missing migration for 0.3.0-0.5.0.
- 0.5.0-0.7.0 was split into 0.5.0-0.6.0 and 0.6.0-0.7.0. It is believed
that the correct migrations are in each one. index.js file was added
and edited as needed. Note the 0.5.0-0.6.0 directory shows all the
files as new even though they were in the previous, larger migration.
- The migration.js was updated for these migrations.
- The enum update for meter_type was changed since it did not
work in a transaction. Note the 0.5.0-0.6.0 language type migration
probably has the same issue but not being fixed since it is assumed
all sites have already done this (and I'm being a little lazy :-).
Note I only tested the 0.6.0-0.7.0 migration since it is tricky to
install 0.5.0 and before.