remove OED_OBVIUS_PASSWORD
OED has an Obvius user role so removing the environment variable to
provide the Obvius password. This should be more secure and also allows
for multiple Obvius user/password if needed.
Note that the Obvius push request needs to provide the email and it does
not seem to do that by default. Thus, the URL needs to be:
OED_Site_URL/api/obvius?email=test@example.com
The password is set by another field in Obvius.
It is believed this will work but has not been tested on a live system.
It is possible that you need to manually provide the & at the end of the
URL above.