fix Obvius test
This makes the test string what is expected. It adds back in the status
check since it now works.

It also includes a change
coming in PR #672 that fixes issues with checking user on Obvius
requests. This will see if it works on GitHub Actions and then be
removed until that PR clears.