rate limiting test script
I took the basic script from @truongdd03 and made it a little more
general. I felt it was good to have it in the testing directory so
others would have access to it. Note this is currently only manual
testing but automating at some point might be good, esp. if we
limit specific routes.