fixes issue with Obvius config files
If a second config file is uploaded with the same meter then it now
checks so it does not try to recreate.
This was causing error msgs during the test suite.
It is unknown if Obvius will do this (probably) and a TODO indicates we need
to decide if the meter should be updated.