shorten dates on graphics
The moment format was changed so it uses fewer characters.
This is a small change but will help make hover and x-axis labels be
shorter so they take up less space.

Note that compare is not done since its labels are being changed.
It is assume that work will do the same changes as this.