fixes bug with compareTimeInterval
OED uses compare times that are to the whole hour since the data is
gotten by the hour/day for comparisons. The code did not initially set
the end time of the compareTimeInterval to the hour. This had two
issues. First, for each new request, the time shifted in the hour and
the Redux state was recalculated to be saved even though it did not
change. This was inefficient. Second, when groups were involved,
the state was to the exact time but the used interval was to the hour
so an error occurred when the value was not found in the state. This
only happened at times but caused compare graphics to not display.