quantity line readings for units added
This is the first step toward allowing unit conversions for graphics.
This commit does:
- Model and DB changes to support Cik storage and using Cik in
DB to create Pik.
- Functions to create Cik and Pik from DB values.
- New DB functions were added to get line readings for graphing.
This was for both meters and groups.
- Migrations for DB changes which begin the ones for next release.
- The commented out code for the minute view was removed since the
raw readings are working fine and this is not expected to be used.

Limitations:
- The actual code for graphics is not yet included.
- This is only for quantity type readings.
- The inserting of Cik into the DB should be a transaction.
- compressed_readings_2 and compressed_group_readings_2 were left
but they are planned to be removed once the conversion to the new
reading functions is complete.