use meter/group then line
In my PR to incorporate units, I chose to use meter/group first and then
line in names since that followed the order in the state. However, a lot
of existing code does it the other way around so this inverts my naming
to be consistent with that. It also uses singular as in the rest of the
code.