fix group line/bar bug
PR #723 fixed this on development. That funcdtion is going
away so the fix is applied here to the new version. Note the same issue
was put into group bar when it was made similar to group line.
It is interesting that group compare already had this fix.

There is also a small fix to remove a variable that was not used.