Get groups mostly working with units
- Changed so when OED gets group details it gets the deep meters.
This happens on startup so that state is always there.
- Remove async from metersInGroup as should not have such a function in
mapStateToProps. This was causing a bug where the lack of an await meant
the groups were processed before the meters were know. Always having
deep meters with groups means you don't need to go to the DB.
- Changed sorting for menu display for groups and made unit be the same.
- Fixed bug in how compatible groups and units for sets determined.
- Added script (not yet fully ready) to add meters and readings for
testing graphing of data with units. Has CSV file with data.
- Added testing functions for Cik and Pik where do not need to pass
a conn.