pipeline understands DST +
This is a good first pass at making the pipeline understand daylight
savings time. It still needs careful review and testing. It also does
not work if the offset is not given in the date/timestamp. It also
is not yet integrated into the full pipeline including CSV uploads.
The major changes are:

- pipeline can deal with DST shifts. See devDoc for more details.
- pipeline changed in how parses date/time strings due to other changes.
The user must now state if want a more relaxed parsing.
- Fixed old pipeline bug where reported non-cumulative readings with
negative value as an error.
- The DB meters table now has the start/end_timestamp to be a string
because that preserves the offset. There is also a new column to
keep the previous end timestamp for when the pipeline has batches of
readings that are separated while processing a cross out of DST.
- When getting meters by id, name use * so always get all columns.
- Tests updated for new meter columns.
- The expected output from the pipeline tests was updated. First, you
get +00:00 not Z for start/end timestamp. Second, the output
when questionable date/time formats are given was changed.
- src/server/test/web/csvPipeline/regAscVariedFormats.csv was changed
to remove the case where the year was not first with hyphens
since the new use of moment does not support this.

Other changes:
- A typo in src/server/models/Meter.js for rul -> url.
- src/server/services/meterTimezone.js changed to use !==.
- src/server/services/csvPipeline/saveCsv.js did not have require
for moment
- src/server/test/db/configfileTests.js checked dates by toISOString()
and changed to moment.format() because had issue where looked at
fractions of a second and it differed.
- src/server/test/web/csvPipelineTest.js used the _i field of moment but
that is not recommended and caused issues. Used format instead.
- src/server/test/web/meters.js fixed for timezone format.
- src/server/test/db/meterTests.js fixed for new way timestamp works
and for default values.