update meter create + more
- Fixed both unit input to display identifier of unit and show sorted.
Also limit to correct type of unit for each.
Admin can still enter bad combinations so more checks
might be desireable in the future.
- Remove some placeholder value for input. Need to consider
others to be consistent.
- Added check on area, duplication & gps. Removed check on input since
gps needs multiple characters to be correct.
- Changed/fixed values not input by user so pass null/undefined to get
default values for meter.
- Make GPS a point for server use.
- Make timestamps be moments for server use or undefined
if not value so get default.
- Cumulative start/end is undefined if no input to get defaults.
- Fixed order of meter create parameters on route.
- Per React directions, removed the defaultValue for value. Should
not mix value with defaultValue (get console warning)
and we use value elsewhere in some cases.
- Change meter create to use MeterEditData type for routing.
- Edits to meter edit, some will be removed/changed as roll
in way done for create.
- Left some TODO for future consideration.
- Misc edits including some formatting