Cik/Pik only contain graphable units
The arrays were including units that could not be graphed. Changed so
get_suffix_type only returns if displayable not none - probably not
important but allows admins to hide created ones. Also changed
getting units of type unit so displayable not none. This excludes
ones originally with a suffix string that OED processes. Before,
they were shown and this was a bug. Finally, updated unit tess
to be consistent with changes.