fix user create bug & possible test log and getConnection issues
There was a missing await in the users create route that was causing the
usersTest.js to fail at times. It appears the issue is now gone.

In the first test in testsTest.js, it does a mocha.before when it uses a
route to log in. If this is the only test run and a previous test left
the DB in an unusual state then there would be no test user in the DB
and this fails. To fix, it calls recreateDB() first. Most other uses are
okay and they have comments added to that effect.

Move so getConnection is before routes are run. Unclear this matters but
all others like this so done. Comment about this in code.

Removed TODO and added correct comment for flow/raw reading tests.