fix pipeline bug where result modified
In one case (end only, shift into DST), the pipeline modified the
endTimestamp. Because moment puts this change in the original object,
it changed the result already put into the array. The software now
clones the Reading before adding to result to avoid this.
Did in every case just to be safe.