replace manual tests with auto ones
This removes all the files associated with the manual tests now that the
automated tests do them. It renames the testing file to be the same as
the old one with limited tests for the pipeline. It also separates
the data files into its own directory to match other tests.