state holds user friendly defaultMeterReadingFrequency
The Redux state now holds defaultMeterReadingFrequency in the user
friendly format rather than the DB format. This removes the hack to
format the value on edit. The routes for get/set preferences does this
so even on edit the state gets what the DB thought the new value was
just in case it differs from what the user set. It also means that the
value shown on page is the standard one and not the format the user
input the value in after save.