fix two meter state issues
- When you edit a meter and make certain changes, it now refreshes
readings so they will graph. The code to refresh now check that either
refresh readings or cik occurred to avoid a window reload when
it is not needed.
- The CSV readings upload now does a window refresh so any newly
created meter is visible. There are TODOs in the code because it
should only happen when the meter is created and it should only
update the meter state rather than reload. This is being delayed until
the page is converted to React hooks.