fix: `peerDependency` resolutions
`react-notification-system` has a bug in which is specifies React 16 as a
`peerDependency`. Since we use React 17, this causes recent versions of `npm`
to error on install. Although this can be partially worked around with `--force`
or `--legacy-peer-deps`, it is only a workaround. Until upstream fixes this, we
can fix this with the `overrides` field. Since React versions are largely
backwards-compatible, we simply set the current version (`$react`) of React (and
`react-dom`) as a peer-dependency.

Similarly, `@formatjs/intl` is not compatible with `typescript@4`. So,
update TypeScript so the `peerDependency` is satisfied.

With these fixes, passing `--legacy-peer-deps` is no longer required.
Yay!