fix: Remove `core-js@2`
When using Babel, `core-js` is a common dependency used for
automatically polyfillying modern ECMAScript features. However, it is a
transitive dependency, so it shouldn't be specified manually as it was. We want to
uninstall it, but it appears other parts of the codebase use its `escapeHtml`
pollyfill. So, replace these uses with the `escape-html` package, then
remove `core-js`.