fix: Make `TimeInterval.mjs` work
The previous `TimeInterval.js` was tricky because it was used on the
server _and_ client. With Vite, the imported `TimeInterval` variables
seemed to always be undefined. It appeared to be a bug with
`TimeInterval.js` being a file with CommonJS exports.

To remedy this, convert the file so it uses ECMAScript modules (thankfully, the
technology is now stable). As a result, imports to this file must be updated as well.