fix: Convert `TimeInterval.js` back to CommonJS
Previously, this file was `TimeInterval.mjs` and exported things with
ESM. This was originally done to make Vite happy, but however, this
didn't actually work server-side. This is because the server is written
in CommonJS - and to import ESM from CommonJS, one needs to use dynamic
import. And, since the imports are at the top of the file, TLA
(top-level await) must be active so the module is implemented
(rather than Promise<Module>, which was previously occuring). And, TLA
is only available in ECMAScript Modules. Migrating everything to ESM would
take a lot of work, and it's not guaranteed everything will work the same.

So revisit keeping `TimeInterval.js` CommonJS. This was previously not
possible becaues Vite errored out. But now, Vite accepts this perfectly
fine.