fix: Vite now listens on port `9229` in container
There seemed to be some issue when Vite was launched as a background
process (with `&` in the shell). When this was done, Vite would not
listen on port 9229; it would detect that port being taken and listen on
9230 instead.

To fix this, modify the docker-compose file so the development server is
given its own service.