fix bug in returned meters info
The newer routes return the meter info on edit and create. They were not
formatting the response so a couple of the properties had the wrong
name. This caused a bug where after creating a meter you could not
edit it until the web browser was refreshed. This should resolve this.