okay version of 3D graphic DB code
This version supports multiple meters for groups. Uses special case for
every hour since much faster. Scales with the number of meters and
the number of points returned. It can take several seconds for 1 year
and every two hours. It probably needs more optimization but should
work reasonably and for first shot. Groups with lots of meters and/or
long times (1 year) may be slow.