bar range fixed if not fully overlapping
If there were multiple meters/groups with different time ranges then
when the user selected a subrange that did not include on of the
meter/group ranges then it did not honor this on bar. This is now
fixed. It now uses all the meters to graph to determine what to do.