remove min/max on line if raw data
The line graphic no longer shows min/max values on raw data. This is
done by setting the values to null in Redux and NaN from the DB.
The graphic shows each meter so if one is raw it will
not show min/max but if another in the same graphic is not raw then
the min/max is shown.