Bug when recalibrate a map
There was a bug that meant after recalibrating a map it did not properly
return to the map overview page. This is now fixed by explicitly going
back to the proper page. Note there is an issue that OED thinks there
was a change to the map and prompts to see if you want to save. This is
being left for now and is noted in the help pages. It may go away when
this page is converted to modals or be fixed at a future time.