centralize testing utils
This partly reporduces work already done on the 3D code but is redone
since 3D not yet ready for PR and my commit of that work mixed
different its so hard to integrate. This may cause merge issues with 3D
and I am willing to help fix that up. This is being done so changes can
be made to the testing code for the next round of tests.
Changes to other testing files will wait for the 3D PR.